OPHJOD-2018: Update profile texts and move welcome path into own file
